refactor: 금액 관련 자료형 수정
금액 관련 자료형 수정 및 payment 엔티티 수정
Close #153

+++ src/main/java/com/example/unbox_be/domain/order/service/OrderServiceImpl.java
@@ -68,7 +68,7 @@ public class OrderServiceImpl implements OrderService {
         User seller = getUserByIdOrThrow(sellingBid.getUserId());
 
         // 5. 가격 타입 변환 (Integer -> BigDecimal)
-        BigDecimal price = BigDecimal.valueOf(sellingBid.getPrice());
+        BigDecimal price = sellingBid.getPrice();
 
         // 6. Order 생성
         Order order = Order.builder()
